refactor: move render-debug and render-stress to standalone binaries
Move from dang subcommands to self-contained cmd/ binaries at
cmd/render-debug and cmd/render-stress. Add q and Ctrl+C as quit
hotkeys for render-stress.
